+++ tests/proud-core-stubs.php
@@ -2,6 +2,10 @@
 
 namespace Proud\Core;
 
+function proud_html_preview_is_resolving_source() {
+	return ! empty( $GLOBALS['filetoweb_test_resolving_source'] );
+}
+
 function proud_html_preview_queue_cleanup( $provider, $artifact_key, $artifact_url ) {
 	$GLOBALS['filetoweb_test_cleanup_queue'][] = array( $provider, $artifact_key, $artifact_url );